# retrace: Fix GPU Duration calculation